requirements.txt
Flask==1.0.2
mysql-connector==2.2.9
dotmap==1.3.13
